+++ src/appConstants/general.constants.ts
@@ -6,6 +6,8 @@ export const ALL_SHARDS_SHARD_ID = 4294967280;
 
 export const TIMEOUT = 10000;
 export const REFRESH_RATE = 6000;
+
+export const SUPERNOVA_REFRESH_RATE = 600;
 export const POOLING_REFRESH_RATE_LIMIT = 1000;
 export const POOLING_REFRESH_RATE = 3000;
 export const LONG_POOLING_REFRESH_RATE = 60 * 1000;