+++ src/redux/selectors/epochProgress.ts
@@ -1,0 +1,9 @@
+import { createSelector } from '@reduxjs/toolkit';
+import { RootState } from '../store';
+
+const stateSelector = (state: RootState) => state.epochProgress;
+
+export const epochRoundsLeftSelector = createSelector(
+  stateSelector,
+  (state) => state.roundsLeft
+);