+++ src/redux/selectors/refresh.ts
@@ -5,7 +5,12 @@ const stateSelector = (state: RootState) => {
   return state.refresh;
 };
 
-export const refreshSelector = createSelector(
+export const refreshTimestampSelector = createSelector(
   stateSelector,
-  (state) => state.refresh
+  (state) => state.refresh.timestamp
+);
+
+export const poolingRefreshTimestampSelector = createSelector(
+  stateSelector,
+  (state) => state.refresh.poolingTimestamp
 );