+++ src/redux/selectors/stats.ts
@@ -6,3 +6,58 @@ const stateSelector = (state: RootState) => {
 };
 
 export const statsSelector = createSelector(stateSelector, (state) => state);
+
+export const statsIsDataReadySelector = createSelector(
+  stateSelector,
+  (state) => state.isDataReady
+);
+
+export const statsIsWebsocketSelector = createSelector(
+  stateSelector,
+  (state) => state.isWebsocket
+);
+
+export const statsEpochSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.epoch
+);
+
+export const statsRefreshRateSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.refreshRate
+);
+
+export const statsRoundsPerEpochSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.roundsPerEpoch
+);
+
+export const statsRoundsPassedSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.roundsPassed
+);
+
+export const statsBlocksSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.blocks
+);
+
+export const statsAccountsSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.accounts
+);
+
+export const statsTransactionsSelector = createSelector(
+  stateSelector,
+  (state) => state.unprocessed.transactions
+);
+
+export const statsEpochPercentageSelector = createSelector(
+  stateSelector,
+  (state) => state.stats.epochPercentage
+);
+
+export const statsEpochTimeRemainingSelector = createSelector(
+  stateSelector,
+  (state) => state.stats.epochTimeRemaining
+);